refactor: moved evaluation methods to src.evaluation.factuality - Added scaffolding for tests
