+++ results/test-extrinsic-10.csv
@@ -1,0 +1,7 @@
+model,total,factual,non_factual,non_factual_extrinsic,non_factual_intrinsic,unknown,skipped,failed,ents_per_sum,sum_with_extrinsic,rouge1,rouge2,rougeL
+fbs_oracle,10,0.8,0.0,0.0,0.0,0.0,0.2,2,2.375,0.7,0.4750156298846559,0.2553732177609856,0.40097719057161707
+fbs_classifier,10,0.5,0.1,0.1,0.0,0.0,0.4,3,3.3333333333333335,0.7,0.4690450348052105,0.23666620118233023,0.38442967559030894
+fbs_classifier_i10,10,0.5,0.1,0.1,0.0,0.0,0.4,4,2.5,0.6,0.48372714378068205,0.24277723471271861,0.3850092246966303
+baseline,10,0.5,0.5,0.5,0.0,0.0,0.0,0,2.7,1.0,0.4878175256184945,0.26047976067766404,0.4296686455130926
+corrector,10,0.5,0.5,0.5,0.0,0.0,0.0,0,2.7,1.0,0.4878175256184945,0.26047976067766404,0.4296686455130926
+gold,10,1.0,0.0,0.0,0.0,0.0,0.0,0,0.0,0.0,1.0,1.0,1.0
