+++ tests/test_evaluation.py
@@ -1,0 +1,6 @@
+from src.evaluation.factuality import evaluate_factuality
+
+# TODO: add tests for sub-routines
+
+def test_evaluate_factuality():
+    pass
